Fixed install.sh to other package managers then apt-get

+++ install.sh
@@ -6,14 +6,17 @@ if [ "$(id -u)" != "0" ]; then
 fi
 
 echo "Adding PWNTools to \$PATH in .bashrc"
-echo "export PATH=$PWD/bin:\$PATH" >> $(getent passwd $SUDO_USER | cut -d: -f6)/.bashrc
+echo "export PATH=$PWD/bin:\$PATH" >> $(getent passwd $SUDO_USER | cut -d: -f6)/.zshrc
 
 
 echo "Adding PWNLib to \$PYTHONPATH in .bashrc"
-echo "export PYTHONPATH=$PWD/lib:\$PYTHONPATH" >> $(getent passwd $SUDO_USER | cut -d: -f6)/.bashrc
+echo "export PYTHONPATH=$PWD/lib:\$PYTHONPATH" >> $(getent passwd $SUDO_USER | cut -d: -f6)/.zshrc
 
 echo "Installing packages"
-DEPS="python-crypto python-paramiko python-sqlalchemy libdistorm64-dev python-gmpy"
-yes | sudo apt-get install ${DEPS}
-
+if [ -f /etc/debian_version ]; then
+    DEPS="python-crypto python-paramiko python-sqlalchemy libdistorm64-dev python-gmpy"
+    yes | sudo apt-get install ${DEPS}
+elif [ -f /etc/arch-release ]; then
+    DEPS="python2-crypto python2-paramiko python2-sqlalchemy"
+    yes | sudo pacman -S 
 echo "All DONE"
